+++ packages/modules/imodel-browser/src/containers/iModelTiles/IModelTile.tsx
@@ -2,7 +2,7 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import { ThemeProvider, Tile } from "@itwin/itwinui-react";
+import { Tile } from "@itwin/itwinui-react";
 import React from "react";
 
 import { ApiOverrides, IModelFull } from "../../types";
@@ -43,7 +43,24 @@ export const IModelTile = ({
   tileProps,
   refetchIModels,
 }: IModelTileProps) => {
-  const moreOptions = React.useMemo(
+  const {
+    name,
+    status,
+    isNew,
+    isLoading,
+    isSelected,
+    thumbnail,
+    badge,
+    leftIcon,
+    rightIcon,
+    buttons,
+    moreOptions,
+    isDisabled,
+    onClick,
+    ...rest
+  } = tileProps ?? {};
+
+  const moreOptionsBuilt = React.useMemo(
     () =>
       _buildManagedContextMenuOptions(
         iModelOptions,
@@ -60,28 +77,47 @@ export const IModelTile = ({
           data: iModel?.thumbnail,
         }
       : undefined;
+
   return (
-    <ThemeProvider theme="inherit">
-      <Tile
-        key={iModel?.id}
-        name={<span title={iModel?.displayName}>{iModel?.displayName}</span>}
-        description={
-          <span title={iModel?.description ?? ""}>
-            {iModel?.description ?? ""}
-          </span>
-        }
-        moreOptions={moreOptions}
-        thumbnail={
-          <IModelThumbnail
-            iModelId={iModel?.id}
-            accessToken={accessToken}
-            onClick={() => onThumbnailClick?.(iModel)}
-            apiOverrides={thumbnailApiOverride}
-          />
-        }
-        className="iac-tile"
-        {...(tileProps ?? {})}
-      />
-    </ThemeProvider>
+    <Tile.Wrapper
+      key={iModel?.id}
+      isNew={isNew}
+      isSelected={isSelected}
+      isLoading={isLoading}
+      status={status}
+      isDisabled={isDisabled}
+      {...rest}
+    >
+      <Tile.Action
+        onClick={(e) => onClick?.(e) ?? onThumbnailClick?.(iModel)}
+        aria-label={`Select iModel ${iModel?.id}`}
+        aria-disabled={isDisabled}
+      >
+        <Tile.ThumbnailArea>
+          {thumbnail ? (
+            <Tile.ThumbnailPicture>{thumbnail}</Tile.ThumbnailPicture>
+          ) : (
+            <IModelThumbnail
+              iModelId={iModel?.id}
+              accessToken={accessToken}
+              apiOverrides={thumbnailApiOverride}
+            />
+          )}
+          {badge && <Tile.BadgeContainer>{badge}</Tile.BadgeContainer>}
+          {leftIcon && <Tile.TypeIndicator>{leftIcon}</Tile.TypeIndicator>}
+          {rightIcon && <Tile.QuickAction>{rightIcon}</Tile.QuickAction>}
+        </Tile.ThumbnailArea>
+      </Tile.Action>
+      <Tile.Name>
+        <Tile.NameLabel>{name ?? iModel?.displayName}</Tile.NameLabel>
+      </Tile.Name>
+      <Tile.ContentArea>
+        <Tile.Description>{iModel?.description ?? ""}</Tile.Description>
+        {(moreOptions || moreOptionsBuilt) && (
+          <Tile.MoreOptions>{moreOptions ?? moreOptionsBuilt}</Tile.MoreOptions>
+        )}
+      </Tile.ContentArea>
+      {buttons && <Tile.Buttons>{buttons}</Tile.Buttons>}
+    </Tile.Wrapper>
   );
 };